feat(Engine): build core file of the Engine
